+++ CHANGELOG.md
@@ -13,6 +13,11 @@
 - `Innmind\OperatingSystem\Sockets::open()` returned socket is now wrapped in a `Innmind\IO\Sockets\Server`
 - `Innmind\OperatingSystem\Sockets::takeOver()` returned socket is now wrapped in a `Innmind\IO\Sockets\Server`
 - `Innmind\OperatingSystem\Ports::open()` returned socket is now wrapped in a `Innmind\IO\Sockets\Server`
+- `Innmind\OperatingSystem\CurrentProcess\Generic::of()` is now declared `internal`
+- `Innmind\OperatingSystem\Filesystem\Generic::of()` is now declared `internal`
+- `Innmind\OperatingSystem\Ports\Unix::of()` is now declared `internal`
+- `Innmind\OperatingSystem\Remote\Generic::of()` is now declared `internal`
+- `Innmind\OperatingSystem\Ports\Sockets::of()` is now declared `internal`
 
 ## 4.2.0 - 2023-12-14